At this point there's no point in having a version of BaseGenerator without the features of BaseObjectiveGenerator.py, so I merged the two together into basegenerator.py and deleted baseobjectivegenerator.py.
